requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-multipart
python-jose[cryptography]